Move all stimulation properties to trigger node
